requirements.txt
fastapi
uvicorn
sqlalchemy
python-telegram-bot==20.6
requests
pandas
openpyxl
psycopg2-binary